需求

+++ book-server/src/main/java/org/cdteam/employee/base/domain/TrainPlanEntity.java
@@ -57,7 +57,7 @@ public class TrainPlanEntity implements Serializable {
     /**
      * 课时
      */
-    private Integer classHour;
+    private Float classHour;
 
     /**
      * 备注
