+++ book-server/src/main/java/org/cdteam/employee/base/dto/TrainPlanCreateDTO.java
@@ -45,7 +45,7 @@ public class TrainPlanCreateDTO {
     /**
      * 课时
      */
-    private Integer classHour;
+    private Float classHour;
 
     /**
      * 备注
