+++ book-server/src/main/java/org/cdteam/employee/base/dto/TrainPlanUploadDTO.java
@@ -34,7 +34,7 @@ public class TrainPlanUploadDTO {
     /**
      * 课时
      */
-    private Integer classHour;
+    private Float classHour;
 
     /**
      * 备注
